LayoutTests/js/function-length.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This tests the length property of functions.");

shouldBe('window.confirm.length', '0');
shouldBe('window.open.length', '2');
shouldBe('window.showModalDialog.length', '1');
shouldBe('window.setTimeout.length', '1');
shouldBe('window.clearTimeout.length', '0');
shouldBe('window.addEventListener.length', '2');

shouldBe('Array.prototype.concat.length', '1');
shouldBe('Array.prototype.join.length', '1');
shouldBe('Array.prototype.push.length', '1');
shouldBe('Array.prototype.slice.length', '2');
shouldBe('Date.UTC.length', '7');
shouldBe('Date.prototype.setSeconds.length', '2');
shouldBe('Date.prototype.setMinutes.length', '3');
shouldBe('Date.prototype.setHours.length', '4');
shouldBe('Date.prototype.setMonth.length', '2');
shouldBe('Date.prototype.setFullYear.length', '3');
shouldBe('Number.prototype.toExponential.length', '1');
shouldBe('Number.prototype.toFixed.length', '1');
shouldBe('Number.prototype.toPrecision.length', '1');
shouldBe('Math.max.length', '2');
shouldBe('Math.min.length', '2');
shouldBe('Function.prototype.apply.length', '2');
shouldBe('Function.prototype.call.length', '1');
shouldBe('String.prototype.split.length', '2');
shouldBe('String.prototype.substring.length', '2');


</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
